% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_pull_orgmetrics_image}
\alias{docker_pull_orgmetrics_image}
\title{Pull the 'orgmetrics' docker container for local use.}
\usage{
docker_pull_orgmetrics_image()
}
\description{
Pull the 'orgmetrics' docker container for local use.
}
\examples{
\dontrun{
docker_pull_orgmetrics_image ()
}
}
