% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{orgmetrics_docker}
\alias{orgmetrics_docker}
\title{Run \code{orgemtrics} in local Docker container}
\usage{
orgmetrics_docker(
  repo_url = NULL,
  github_name = NULL,
  git_email = NULL,
  aggregation_period = 365,
  quarto_local = TRUE,
  quarto_publish = TRUE,
  quarto_provider = "gh-pages"
)
}
\arguments{
\item{repo_url}{URL of GitHub repository of R-universe-like repository
containing a "packages.json" file. If not specified, the Git remote URL of
the repository in the current working directory will be used.}

\item{github_name}{Your username on GitHub. If not specified, value will be
taken from \pkg{whoami}.}

\item{git_email}{Your email used in Git commits. If not specified, value
will be taken from \pkg{whoami}.}

\item{quarto_local}{If \code{TRUE} (default), the Docker command will create a
\verb{quarto/} sub-directory in the directory where this function is called.
This may be used in conjunction with \code{quarto_publish = FALSE} to create local
\verb{quarto/} sub-directory only, without publishing. The contents of the
resultant directory can then be edited or changed in any way, and \verb{quarto publish} run locally there, to enable complete control over the process.}

\item{quarto_publish}{If \code{TRUE} (default), publish dashboard directly to the
nominated Quarto service provider.}

\item{quarto_provider}{Provider of service where dashboard is to be
published.}

\item{aggregation}{period Time period in days over which data are to be
aggregated. Should generally be left at default value of annual aggregation.}
}
\value{
Nothing.
}
\description{
Downloads the 'orgmetrics' docker container and uses that to build a
dashboard of your r-universe-like GitHub organization.
}
\examples{
\dontrun{
# First pull the Docker image
docker_pull_orgmetrics_image ()

# Run the container to build your dashboard locally
orgmetrics_docker (
    repo_url = "https://github.com/your-org/your-r-universe-repo",
    github_name = "your-username",
    git_email = "your-email@example.com"
)
}

}
