% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgmetricsDashboard-package.R
\docType{package}
\name{orgmetricsDashboard-package}
\alias{orgmetricsDashboard}
\alias{orgmetricsDashboard-package}
\title{orgmetricsDashboard: Dashboard of Metrics for Your GitHub Organization}
\description{
Dashboard of metrics for your GitHub organization. Call one function to generate an interactive dashboard displaying the state of your organization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci-review-tools/orgmetricsDashboard}
  \item Report bugs at \url{https://github.com/ropensci-review-tools/orgmetricsDashboard/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

}
\keyword{internal}
