% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gha.R
\name{use_github_action_orgmetrics}
\alias{use_github_action_orgmetrics}
\title{Use orgemtrics Github Action}
\usage{
use_github_action_orgmetrics(
  dir = ".github/workflows",
  overwrite = FALSE,
  file_name = "orgmetrics.yaml"
)
}
\arguments{
\item{dir}{Directory the file is written to.}

\item{overwrite}{Overwrite existing file?}

\item{file_name}{Name of the workflow file.}
}
\value{
The path to the new file, invisibly.
}
\description{
Creates a Github workflow file in \code{dir} to automatically deploy your
'orgmetrics' dashboard.
}
